+++ src/main/java/com/cburch/logisim/gui/prefs/WindowOptions.java
@@ -37,6 +37,7 @@ class WindowOptions extends OptionsPanel {
   private static final long serialVersionUID = 1L;
   private final PrefBoolean[] checks;
   private final PrefOptionList toolbarPlacement;
+  private final PrefOptionList canvasPlacement;
   private final ZoomSlider zoomValue;
   private final JLabel lookfeelLabel;
   private final JLabel zoomLabel;
@@ -63,12 +64,22 @@ class WindowOptions extends OptionsPanel {
   public WindowOptions(PreferencesFrame window) {
     super(window);
 
-    SettingsChangeListener listener = new SettingsChangeListener();
+    final var listener = new SettingsChangeListener();
+    final var panel = new JPanel(new TableLayout(2));
 
     checks =
         new PrefBoolean[] {
           new PrefBoolean(AppPreferences.SHOW_TICK_RATE, S.getter("windowTickRate")),
         };
+
+    canvasPlacement =
+        new PrefOptionList(
+            AppPreferences.CANVAS_PLACEMENT,
+            S.getter("windowCanvasLocation"),
+            new PrefOption[] {
+              new PrefOption(Direction.EAST.toString(), Direction.EAST.getDisplayGetter()),
+              new PrefOption(Direction.WEST.toString(), Direction.WEST.getDisplayGetter())
+            });
 
     toolbarPlacement =
         new PrefOptionList(
@@ -82,10 +93,11 @@ class WindowOptions extends OptionsPanel {
               new PrefOption(AppPreferences.TOOLBAR_HIDDEN, S.getter("windowToolbarHidden"))
             });
 
-    final var panel = new JPanel(new TableLayout(2));
+    panel.add(canvasPlacement.getJLabel());
+    panel.add(canvasPlacement.getJComboBox());
+
     panel.add(toolbarPlacement.getJLabel());
     panel.add(toolbarPlacement.getJComboBox());
-
 
     canvasBgColorTitle = new JLabel(S.get("windowCanvasBgColor"));
     canvasBgColor = new ColorChooserButton(window, AppPreferences.CANVAS_BG_COLOR);
@@ -159,7 +171,7 @@ class WindowOptions extends OptionsPanel {
     setLayout(new TableLayout(1));
     final var but = new JButton();
     but.addActionListener(listener);
-    but.setActionCommand("reset");
+    but.setActionCommand(cmdResetWindowLayout);
     but.setText(S.get("windowToolbarReset"));
     add(but);
     for (final var check : checks) {