+++ src/main/java/com/cburch/logisim/util/LocaleSelector.java
@@ -43,11 +43,9 @@ class LocaleSelector extends JList implements LocaleListener, ListSelectionListe
     }
 
     void update(Locale current) {
-      if (current != null && current.equals(locale)) {
-        text = locale.getDisplayName(locale);
-      } else {
-        text = locale.getDisplayName(locale) + " / " + locale.getDisplayName(current);
-      }
+      text = (current != null && current.equals(locale))
+              ? locale.getDisplayName(locale)
+              : locale.getDisplayName(locale) + " / " + locale.getDisplayName(current);
     }
   }
 
@@ -58,7 +56,7 @@ class LocaleSelector extends JList implements LocaleListener, ListSelectionListe
   @SuppressWarnings("unchecked")
   LocaleSelector(Locale[] locales) {
     setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
-    DefaultListModel<LocaleOption> model = new DefaultListModel<>();
+    final var model = new DefaultListModel<LocaleOption>();
     items = new LocaleOption[locales.length];
     for (var i = 0; i < locales.length; i++) {
       items[i] = new LocaleOption(locales[i]);