+++ src/test/java/com/cburch/logisim/TestBase.java
@@ -9,18 +9,19 @@
 
 package com.cburch.logisim;
 
+import java.util.List;
 import java.util.Random;
 
 public class TestBase {
 
-  public static final int DEFAULT_RANDOM_STRING_LENGTH = 16;
+  public static final int DEFAULT_RANDOM_STRING_LENGTH_MIN = 16;
+  public static final int DEFAULT_RANDOM_STRING_LENGTH_MAX = 32;
 
   /**
    * Function to generate a random string of specified length.
    *
    * @param length Numbers of characters to be generated.
    * @param includeDigits If True, returned string will also contain digits
-   *
    * @return Random string.
    */
   protected String getRandomString(int length, boolean includeDigits) {
@@ -41,7 +42,8 @@ public class TestBase {
   }
 
   protected String getRandomString(boolean includeDigits) {
-    return getRandomString(DEFAULT_RANDOM_STRING_LENGTH, includeDigits);
+    final var len = getRandomInt(DEFAULT_RANDOM_STRING_LENGTH_MIN, DEFAULT_RANDOM_STRING_LENGTH_MAX);
+    return getRandomString(len, includeDigits);
   }
 
   protected String getRandomString(int length) {
@@ -49,11 +51,19 @@ public class TestBase {
   }
 
   protected String getRandomString() {
-    return getRandomString(DEFAULT_RANDOM_STRING_LENGTH);
+    final var len = getRandomInt(DEFAULT_RANDOM_STRING_LENGTH_MIN, DEFAULT_RANDOM_STRING_LENGTH_MAX);
+    return getRandomString(len);
   }
 
   protected static int getRandomInt(int min, int max) {
     if (min >= max) throw new IllegalArgumentException("Max must be greater than Min.");
     return (new Random()).nextInt((max - min) + 1) + min;
   }
-}
+
+  /** Dumps content of given list to stdout. */
+  protected void printList(List<String> list) {
+    for (final var line : list) {
+      System.out.println(line);
+    }
+  }
+} // End of TestBase